requirements.txt
wheel
sphinx >= 3.0
sphinx_rtd_theme
sphinxcontrib-trimblank
pyyaml
jsonschema < 4.18.0
GitPython
sphinx-lint
google-auth-oauthlib
google-api-python-client
pytz
tools/yaml2x/get_yaml_data
